requirements.txt
# Discord 资源分发 Bot
discord.py>=2.0
python-dotenv>=1.0.0
